+++ include/flexisip/module-router.hh
@@ -103,6 +103,7 @@ protected:
 	bool mFallbackParentDomain{};
 	bool mAllowDomainRegistrations{};
 	bool mEnableCallDiversions{};
+	int mNoContactForAorReturnCode{};
 
 private:
 	static ModuleInfo<ModuleRouter> sInfo;