+++ src/modules/module-router.cc
@@ -273,6 +273,13 @@ void ModuleRouter::declareConfig(GenericStruct& moduleConfig) {
 	        "This prevents loops during successive call diversions.",
 	        "10",
 	    },
+	    {
+	        Integer,
+	        "no-contact-for-aor-return-code",
+	        "SIP response code to send when the requested AoR exists but has no valid registered contact.\n"
+	        "Valid interval: [400; 599]",
+	        "404",
+	    },
 	    config_item_end,
 	};
 	moduleConfig.addChildrenValues(configs);
@@ -331,6 +338,12 @@ void ModuleRouter::onLoad(const GenericStruct* mc) {
 	}
 
 	mVoicemailServerUri = SipUri{mc->get<ConfigString>("voicemail-server")->read()};
+
+	const auto* noDevicesRegisteredReturnCodeParam = mc->get<ConfigInt>("no-contact-for-aor-return-code");
+	mNoContactForAorReturnCode = noDevicesRegisteredReturnCodeParam->read();
+	if (mNoContactForAorReturnCode < 400 || mNoContactForAorReturnCode > 599) {
+		throw BadConfigurationValue{noDevicesRegisteredReturnCodeParam};
+	}
 }
 
 void ModuleRouter::sendReply(RequestSipEvent& ev, int code, const char* reason, int warn_code, const char* warning) {
@@ -426,7 +439,7 @@ void ModuleRouter::routeRequest(unique_ptr<RequestSipEvent>&& ev, const shared_p
 
 				mForkManager->fork(std::move(ev), sipUri, forkContacts, mDomains);
 			} else {
-				sendReply(*ev, SIP_404_NOT_FOUND);
+				sendReply(*ev, mNoContactForAorReturnCode, sip_status_phrase(mNoContactForAorReturnCode));
 			}
 		}
 		return;